Add cached statistics to front page for played games

+++ web/model/game_stat.go
@@ -5,6 +5,6 @@ import (
 )
 
 type GameStatistic struct {
-	MinutesPlayed int
+	MinutesPlayed int64
 	Game          sharedModel.Game
 }
